fix: Change updated_at and created_at attributes to use datetime.now()
This commit fixes an issue where the `updated_at` and `created_at` attributes of our models were not being set to the correct date and time.

To fix the issue, we're now using the `datetime.datetime.now()` function to set the attributes. This function returns the current date and time, which will ensure that our attributes are always up-to-date.